% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greg2jal.R
\name{greg2jal0}
\alias{greg2jal0}
\title{Convert Gregorian calendar date to Jalali calendar date}
\usage{
greg2jal0(year, month, day)
}
\arguments{
\item{year}{An integer specifying Gregorian year}

\item{month}{An integer specifying Gregorian month}

\item{day}{An integer specifying Gregorian day}
}
\value{
An integer vector consisting of the corresponding Jalali year, month and day
}
\description{
Convert Gregorian calendar date to Jalali calendar date
}
\examples{
greg2jal0(622, 3, 21)
greg2jal0(1983, 9, 8)

}
\author{
Abdollah Jalilian
}
