% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isLeap.R
\name{isLeap}
\alias{isLeap}
\title{Determining whether a given year is leap year}
\usage{
isLeap(year, cal = "Gregorian")
}
\arguments{
\item{year}{An integer vector specifying given years}

\item{cal}{A character string specifying the calender type. Only 'Gregorian' and 'Jalali' are implemented}
}
\value{
A logical vector of of the same length as \code{year} which indicates wheter given years are leap years or not
}
\description{
Determining whether a given year is leap year
}
\examples{
isLeap(1362, 'Jalali')
isLeap(c(2000, 2100))

}
\author{
Abdollah Jalilian
}
