% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greg2jal.R
\name{greg2jal}
\alias{greg2jal}
\title{Convert Gregorian calendar date to Jalali calendar date}
\usage{
greg2jal(date)
}
\arguments{
\item{date}{An object of class Date specifying Gregorian date}
}
\value{
An integer vector consisting of the corresponding Jalali year, month and day
}
\description{
Convert Gregorian calendar date to Jalali calendar date
}
\examples{
greg2jal(base::Sys.Date())

}
\author{
Abdollah Jalilian
}
