% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jal2greg.R
\name{jal2greg}
\alias{jal2greg}
\title{Convert Jalali calendar date to Gregorian calendar date}
\usage{
jal2greg(year, month, day, asDate = TRUE)
}
\arguments{
\item{year}{An integer specifying Jalali year}

\item{month}{An integer specifying Jalali month}

\item{day}{An integer specifying Jalali day}

\item{asDate}{A logical flag indicating whether the output Gregorian date must be in date format}
}
\value{
If \code{asDate = TRUE}, the default case, an object of the \code{Date} class in \code{R}, otherwise an integer vector consisting of the Gregorian year, month and day
}
\description{
Convert Jalali calendar date to Gregorian calendar date
}
\examples{
jal2greg(1, 1, 1)
jal2greg(1362, 6, 17)
jal2greg(1362, 6, 17, asDate=FALSE)

}
\author{
Abdollah Jalilian
}
